ingestion_api/requirements.txt
azure-core==1.38.2
azure-functions==1.24.0
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
idna==3.11
isodate==0.7.2
MarkupSafe==3.0.3
py4j==0.10.9.9
pycparser==3.0
pyspark==4.1.1
python-dotenv==1.2.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
